Move bom types to submodule (#703)
