+++ doc/changelog.d/703.documentation.md
@@ -1,0 +1,1 @@
+Move bom types to submodule
