+++ doc/source/api/bom_builder/eco2301.rst
@@ -1,32 +1,33 @@
-.. _ref_grantami_bomanalytics_bom_api:
+.. _ref_grantami_bomanalytics_bom_eco2301:
 
-BoM object model
-================
+23/01 XML BoM types
+===================
+
+.. currentmodule:: ansys.grantami.bomanalytics.bom_types.eco2301
 
 The models documented in this section are Python bindings for the Ansys Granta MI 23/01 XML schema.
 Not all elements are required to define a valid BoM for analysis, and not all elements have an impact
 on all types of analysis. For information on the relevant items for each analysis type,
 see :ref:`ref_grantami_bomanalytics_api_index`.
 
+Classes on this page should be imported from the module ``ansys.grantami.bomanalytics.bom_types.eco2301``.
+
 .. note::
    The following elements are defined in the 23/01 BoM schema but are not supported in this module:
 
-   * ``NonMIPartReference`` on :class:`~ansys.grantami.bomanalytics.bom_types._bom_types.Part`
-   * ``Annotations`` and ``AnnotationSources`` on :class:`~ansys.grantami.bomanalytics.bom_types._bom_types.BillOfMaterials`
+   * ``NonMIPartReference`` on :class:`~.Part`
+   * ``Annotations`` and ``AnnotationSources`` on :class:`~.BillOfMaterials`
 
    This module does not contain classes that correspond to these types and cannot serialize a BoM that includes these
    elements. It is still possible to deserialize an XML BoM that uses these elements, but these elements cannot be
    converted to Python objects.
 
-.. versionadded:: 2.0
+.. versionchanged:: 2.3
+   The types documented on this page have moved from ``ansys.grantami.bomanalytics.bom_types`` to
+   ``ansys.grantami.bomanalytics.bom_types.eco2301``. They are still importable from the old location for
+   compatibility.
 
-.. _ref_grantami_bomanalytics_api_billofmaterials:
-.. autoclass:: ansys.grantami.bomanalytics.bom_types._bom_types.BillOfMaterials
+.. automodule:: ansys.grantami.bomanalytics.bom_types.eco2301
    :inherited-members:
    :member-order: bysource
-
-
-.. automodule:: ansys.grantami.bomanalytics.bom_types._bom_types
-   :exclude-members: BillOfMaterials, BaseType, HasNamespace, SupportsCustomFields
-   :inherited-members:
-   :member-order: bysource
+   :exclude-members: BoMReader, BoMWriter
