+++ doc/source/api/bom_builder/gbt1205.rst
@@ -1,0 +1,22 @@
+.. _ref_grantami_bomanalytics_bom_gbt1205:
+
+12/05 Granta Base Types
+=======================
+
+.. currentmodule:: ansys.grantami.bomanalytics.bom_types.gbt1205
+
+The models documented in this section are Python bindings for the Granta Base Types 12/05 XML schema.
+These are typically used to define references to attributes and records in a Granta MI database.
+
+Classes on this page should be imported from the module ``ansys.grantami.bomanalytics.bom_types.gbt1205``.
+
+Alternatively, use the :ref:`ref_grantami_bomanalytics_bom_builders`.
+
+.. versionchanged:: 2.3
+   The types documented on this page have moved from ``ansys.grantami.bomanalytics.bom_types`` to
+   ``ansys.grantami.bomanalytics.bom_types.gbt1205``. They are still importable from the old location for
+   compatibility.
+
+.. automodule:: ansys.grantami.bomanalytics.bom_types.gbt1205
+   :inherited-members:
+   :member-order: bysource
