+++ src/ansys/grantami/bomanalytics/bom_types/eco2301/__init__.py
@@ -20,8 +20,26 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from ._builders import AttributeReferenceBuilder, RecordReferenceBuilder
-from .eco2301 import *
-from .eco2301._bom_reader import BoMReader
-from .eco2301._bom_writer import BoMWriter
-from .gbt1205 import *
+from ._bom_types import *
+
+__all__ = [
+    "BillOfMaterials",
+    "Part",
+    "Material",
+    "Process",
+    "Substance",
+    "Specification",
+    "TransportStage",
+    "BoMDetails",
+    "UsePhase",
+    "ProductLifeSpan",
+    "UtilitySpecification",
+    "StaticMode",
+    "MobileMode",
+    "ElectricityMix",
+    "Location",
+    "UnittedValue",
+    "EndOfLifeFate",
+    "Category",
+    "DimensionType",
+]
